Support BYO_CRYPTO by allowing tls_ctx to be passed into providers.
To support BYO_CRYPTO, every credentials provider needs to let the user pass in their own TLS context.

The default-chain and ECS providers did not allow this. Now they do.

The STS and STS-web-identity providers were creating a TLS context if the user was lazy and didn't pass one in. This is a bad idea for several reasons so now the user is required to pass one in. This api-breaking change is ok because we haven't bound these classes anywhere yet.

For backwards compatibility, the default-chain provider will create a TLS context if one is not supplied (it's required if BYO_CRYPTO is defined)