default chain provider: plug in STS web identity provider
In clusters with temporary credentials, the default chain provider will only
query the ECS and IMDS proiders, but not the STS web identity provider.

This differs from the use-case underlying the `DefaultAWSCredentialsProviderChain`
in use by the AWSK SDK, which does check the STS assume-role-with-web-identity
provider before checking the ECS and IMDS credential providers.

For clusters with temporary STS credentials, the default chain provider needs
such support, hence plug in the STS web identity provider.