Additional logging for default credentials chain
We've been finding it difficult to parse the log output of credentials
providers when investigating permissions issues. The problem is that the
default chain is constructed somewhat dynamically, with some providers
only added conditionally, and some dispatching differently depending on
environment variables. This patch just adds some more logging in the
default chain and in the providers it uses, so we can more easily trace
the path of credentials resolution.

Signed-off-by: James Bornholt <bornholt@amazon.com>