When sourcing credentials from an external process, ignore stderr (#250)
**Issue:**
If the external process logged to `stderr` during a normal successful run, the credentials would fail to parse.

This was happening because the credentials-provider would always combine `stderr` and `stdout` by appending `2>&1` to the external command. Then JSON parsing would fail, due to random lines of logging on `stderr` mixing with valid JSON on `stdout`.

**Description of changes:**
Instead of redirecting `stderr` to `stdout`, redirect it to `/dev/null`.

It would be better to capture `stderr` separately, and display it if the external process fails. But `aws_process_run()` doesn't currently capture `stderr`, and would require a rework to do so.